Adding functionnality, correcting some issues + some optimization
Adding the info in volumePage, still need to do the Contain.
Correcting the issue with the MyPictureDisplay.
Correcting the fact the clicking on the go back button make you go back to /search instead of where you where before.
Optimizing a bit the volumePage

+++ lib/src/const/own_icon.dart
@@ -4,13 +4,9 @@ import 'package:flutter_svg/flutter_svg.dart';
 class OwnIcon extends StatelessWidget {
   Color iconColor;
   final String iconName;
-  final int? height;
+  final double? height;
 
-  OwnIcon(
-      {required this.iconColor,
-      required this.iconName,
-      this.height,
-      super.key});
+  OwnIcon({required this.iconColor, required this.iconName, this.height, super.key});
 
   @override
   Widget build(BuildContext context) {
